chore: move schemas to shared
