+++ shared/schemas/article.ts
