+++ shared/schemas/event.ts
