+++ shared/schemas/planning.ts
