+++ src/views/Assignments/AssignmentsList.tsx
@@ -4,8 +4,8 @@ import type { ColumnDef } from '@tanstack/react-table'
 import { Error as ErrorView } from '../Error'
 import { useDocuments } from '@/hooks/index/useDocuments'
 import { constructQuery } from '@/hooks/index/useDocuments/queries/views/assignments'
-import { fields } from '@/hooks/index/useDocuments/schemas/assignments'
-import type { Assignment, AssignmentFields } from '@/hooks/index/useDocuments/schemas/assignments'
+import { fields } from '@/shared/schemas/assignments'
+import type { Assignment, AssignmentFields } from '@/shared/schemas/assignments'
 import { getUTCDateRange } from '@/shared/datetime'
 import { SortingV1 } from '@ttab/elephant-api/index'
