+++ src/views/EventsOverview/index.tsx
@@ -8,7 +8,7 @@ import { EventsList } from './EventsList'
 import { Header } from '@/components/Header'
 import { Commands } from '@/components/Commands'
 import { eventTableColumns } from './EventsListColumns'
-import { type Event } from '@/hooks/index/useDocuments/schemas/event'
+import { type Event } from '@/shared/schemas/event'
 import { useSections } from '@/hooks/useSections'
 import { useQuery } from '@/hooks/useQuery'
 import { useOrganisers } from '@/hooks/useOrganisers'
