+++ src/views/PlanningOverview/PlanningList.tsx
@@ -5,8 +5,8 @@ import { useCallback, useMemo } from 'react'
 import { Table } from '@/components/Table'
 import { useDocuments } from '@/hooks/index/useDocuments'
 import type { ColumnDef } from '@tanstack/react-table'
-import type { Planning, PlanningFields } from '@/hooks/index/useDocuments/schemas/planning'
-import { fields } from '@/hooks/index/useDocuments/schemas/planning'
+import type { Planning, PlanningFields } from '@/shared/schemas/planning'
+import { fields } from '@/shared/schemas/planning'
 import { SortingV1 } from '@ttab/elephant-api/index'
 import { constructQuery } from '@/hooks/index/useDocuments/queries/views/plannings'
 import { useQuery } from '@/hooks/useQuery'
