+++ src/views/PrintArticles/PrintFlows.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react'
 import { useDocuments } from '@/hooks/index/useDocuments'
 import { useRegistry } from '@/hooks/useRegistry'
-import type { PrintFlow, PrintFlowFields } from '@/hooks/index/useDocuments/schemas/printFlow'
+import type { PrintFlow, PrintFlowFields } from '@/shared/schemas/printFlow'
 import { Button, Select, SelectContent, SelectItem, SelectTrigger } from '@ttab/elephant-ui'
-import { fields } from '@/hooks/index/useDocuments/schemas/printFlow'
+import { fields } from '@/shared/schemas/printFlow'
 import { toast } from 'sonner'
 import { ViewHeader } from '@/components/View/ViewHeader'
 import { View } from '@/components/View'
