+++ src/views/SearchOverview/SearchResult.tsx
@@ -6,9 +6,9 @@ import { LoadingText } from '@/components/LoadingText'
 import { Toolbar } from './Toolbar'
 import { useOrganisers } from '@/hooks/useOrganisers'
 import { useAuthors } from '@/hooks/useAuthors'
-import type { Event, EventFields } from '@/hooks/index/useDocuments/schemas/event'
-import type { Planning, PlanningFields } from '@/hooks/index/useDocuments/schemas/planning'
-import type { Article, ArticleFields } from '@/hooks/index/useDocuments/schemas/article'
+import type { Event, EventFields } from '@/shared/schemas/event'
+import type { Planning, PlanningFields } from '@/shared/schemas/planning'
+import type { Article, ArticleFields } from '@/shared/schemas/article'
 import { useDocuments } from '@/hooks/index/useDocuments'
 import { createSearchColumns } from './lib/createSearchColumns'
 import type { SearchKeys } from '@/hooks/index/useDocuments/queries/views/search'
