+++ src/views/SearchOverview/lib/articleColumns.tsx
@@ -5,7 +5,7 @@ import { DocumentStatus } from '@/components/Table/Items/DocumentStatus'
 import { Title } from '@/components/Table/Items/Title'
 import { NewsvalueMap } from '@/defaults/newsvalueMap'
 import { Newsvalues } from '@/defaults/newsvalues'
-import type { Article } from '@/hooks/index/useDocuments/schemas/article'
+import type { Article } from '@/shared/schemas/article'
 import type { Column, ColumnDef } from '@tanstack/react-table'
 import { DocumentStatuses } from '@/defaults/documentStatuses'
 import { CircleCheck, Pen, Shapes, SignalHigh } from '@ttab/elephant-ui/icons'
