+++ src/views/Wires/WiresList.tsx
@@ -2,8 +2,8 @@ import { useCallback } from 'react'
 import { useQuery } from '@/hooks'
 
 import { Table } from '@/components/Table'
-import type { WireFields } from '@/hooks/index/useDocuments/schemas/wire'
-import { fields, type Wire } from '@/hooks/index/useDocuments/schemas/wire'
+import type { WireFields } from '@/shared/schemas/wire'
+import { fields, type Wire } from '@/shared/schemas/wire'
 import type { ColumnDef } from '@tanstack/react-table'
 import { Toolbar } from './Toolbar'
 import { useDocuments } from '@/hooks/index/useDocuments'
