+++ src/views/Wires/components/PreviewSheet.tsx
@@ -5,7 +5,7 @@ import { FaroErrorBoundary } from '@grafana/faro-react'
 import { Error, Wire } from '@/views'
 import { useNavigationKeys } from '@/hooks/useNavigationKeys'
 import { useModal } from '@/components/Modal/useModal'
-import type { Wire as WireType } from '@/hooks/index/useDocuments/schemas/wire'
+import type { Wire as WireType } from '@/shared/schemas/wire'
 import type { Status as DocumentStatuses } from '@ttab/elephant-api/repository'
 import { MetaSheet } from '@/views/Editor/components/MetaSheet'
 import { useEffect, useMemo, useRef } from 'react'
