+++ src/views/Wires/index.tsx
@@ -15,7 +15,7 @@ import { ViewDialogClose } from '@/components/View/ViewHeader/ViewDialogClose'
 import { ViewFocus } from '@/components/View/ViewHeader/ViewFocus'
 import { Button } from '@ttab/elephant-ui'
 import { useUserTracker } from '@/hooks/useUserTracker'
-import type { Wire } from '@/hooks/index/useDocuments/schemas/wire'
+import type { Wire } from '@/shared/schemas/wire'
 const BASE_URL = import.meta.env.BASE_URL
 
 const meta: ViewMetadata = {
